+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_05_12_084817) do
+ActiveRecord::Schema.define(version: 2021_06_02_082901) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -337,6 +337,10 @@ ActiveRecord::Schema.define(version: 2021_05_12_084817) do
     t.string "system_email_copy"
     t.string "email_from"
     t.string "google_analytics_tracking_code_v4"
+    t.text "header_message"
+    t.boolean "header_message_published", default: false
+    t.datetime "header_message_published_from"
+    t.datetime "header_message_published_until"
     t.index ["domain"], name: "index_folio_sites_on_domain"
   end